My inventory patch everyone's been testing
I'm still wary, but let's toss it into our release, shall we?
This changes the maps in inventorymodel to unordered maps.
On the surface this seems like a good thing, definitely is for inventory loading.
It seemed a lot snappier to literally everyone before the HTTP Asset patch rolled out, so that's good.

Thanks to Router, Bitten, Gooz, Deltek, Damian, Melaniecosti, Torric, and anyone else I might be forgetting at the moment who tested this for me!