Feature Request: Trim spaces off of end of script lines
In order to preserve your ability to edit, this only happens to the asset
we send to the server, you will not notice until you reopen the script.
This is done as optimally as possible, upon each save.
This feature takes into account the possibility that spaces may
be following a quote and therefore it does not trim spaces in these cases.
It is also aware that you may use escaped quotes inside a string, and that
that does not mark the end of a string and therefore whitespace
is not stripped then.