What happens to a ShowParcelOwners deferred?
Does it render like a ShowParcelOwners in the forward?
Or fester like a bug when it's run?
Does it get reported like rotten meat?
Or forgot and coded over
giving users a treat?

Maybe it just sags like a heavy bird...
Prepare to Implode!

Thanks to Beware Hax for reporting this